made software-properties-common install conditional

+++ moloch-setup.sh
@@ -47,8 +47,10 @@ execute n $command
 command='diff /tmp/ethtool1 /tmp/ethtool2'
 execute n $command
 
-command='time apt-get -y install software-properties-common'
-execute $command
+if ! dpkg -l | grep -q software-properties-common; then
+  command='time apt-get -y install software-properties-common'
+  execute $command
+fi
 if ! dpkg -l | grep -q oracle-java8-installer; then
   command='echo oracle-java8-installer shared/accepted-oracle-license-v1-1 select true | /usr/bin/debconf-set-selections -v'
   execute $command
